feat: move validationSchema back up to CTA
