+++ apps/badass/src/components/call-to-action-form.tsx
@@ -5,8 +5,16 @@ import {
   redirectUrlBuilder,
   SubscribeToConvertkitForm,
 } from '@skillrecordings/skill-lesson/convertkit'
+import * as Yup from 'yup'
 
 import LevelUp from '../../public/assets/level-up@2x.png'
+
+const validationSchema = Yup.object().shape({
+  email: Yup.string().email('Invalid email address').required('Required'),
+  first_name: Yup.string()
+    .min(2, 'Name must be at least 2 characters')
+    .required('Required'),
+})
 
 export const CallToActionForm: React.FC<React.PropsWithChildren<any>> = ({
   content,
@@ -46,6 +54,7 @@ export const CallToActionForm: React.FC<React.PropsWithChildren<any>> = ({
                     router.push(redirectUrl)
                   }
                 }}
+                validationSchema={validationSchema}
                 validateOnChange={true}
               />
               <p className="text-center md:text-left lg:text-center text-base text-badass-gray-300 leading-tight mt-6 md:mt-8">
